requirements.txt
click
gql[httpx]
gtfs-realtime-bindings
mako
polyline
pyyaml
requests
shapely